fix: 修复发现页面数据解析错误问题

+++ lib_api/src/main/java/com/imooc/lib_api/model/playlist/MainRecommendPlayListBean.java
@@ -71,7 +71,7 @@ public class MainRecommendPlayListBean {
         private String name;
         private String copywriter;
         private String picUrl;
-        private int playcount;
+        private long playcount;
         private long createTime;
         private CreatorBean creator;
         private int trackCount;
@@ -118,11 +118,11 @@ public class MainRecommendPlayListBean {
             this.picUrl = picUrl;
         }
 
-        public int getPlaycount() {
+        public long getPlaycount() {
             return playcount;
         }
 
-        public void setPlaycount(int playcount) {
+        public void setPlaycount(long playcount) {
             this.playcount = playcount;
         }
